add script to check training data
lotaru only reads the first six records per workflow and task as
training data. That is because only traces from the first six workflow
executions should be used as training data. But since one task could be
executed multiple times per workflow we have to make sure that that is
not the case.